Creates failing test for new feature: caching non-zend files
In our presentation scenario, Bob's boss wants him to add a new feature. The company is starting to use other frameworks in different projects based on individual strengths and requirements, and wants to be able to leverage this library for different framework namespaces in addition to Zend. We start this feature by creating a failing test, imagining that our Cacher namespace is a framework namespace which we want to cache.